feat: seperate reducer list
