+++ web/package.json
@@ -10,14 +10,18 @@
     "@types/node": "^16.18.12",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
+    "@types/react-redux": "^7.1.25",
     "axios": "^1.3.4",
     "babel-jest": "^29.5.0",
     "jest-mock": "^29.5.0",
     "pace-js": "^1.2.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
     "react-router-dom": "^6.8.2",
     "react-scripts": "5.0.1",
+    "redux": "^4.2.1",
+    "redux-typescript": "^1.2.1",
     "styled-components": "^5.3.8",
     "web-vitals": "^2.1.4"
   },
